requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.2
django-crispy-forms==2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
mysqlclient==2.2.6
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
